OWRT-7129: add patches to support multiple DSA CPU ports.
SmartOS won't ever have multiple CPU ports active for the time
being.. this is needed so we can set the DSA CPU port to the
MT7531 RGMII port instead of SGMII as used on the 841-t6 w/MT7531.

We could theoretically use this on the 834-5 but there's no valid
case where the single SGMII link could become saturated so adds an
unnecessary level of complexity.