OWRT-8250: hack to reset GPY211 rate adaptation after upgrade
If upgrading from a build with rate adaptation enabled on the GPY211 to one
with it disabled, the phy can get stuck in a broken state on platforms which
don't have a working phy reset line. In the case of SmartOS, the only target
this affects is the 854 where the GPY211 is connected to MT7531 port 5. The
only way to detect this condition is presence of RX errors but no RX packets.

If the failure condition is detected it can usually be cleared by toggling the
rate adaptation bit once or in some very rare cases twice.