SDG-4420: net: ethernet: mediatek: fall back to random MAC
fall back to using a random MAC if one can’t be found via OF. for
Adtran hardware, mfginfo should always be available on production
units via nvmem. for development boards we have an initramfs
image which formats the flash and loads SmartOS.. during this
stage we need network access but mfginfo is not populated.